+++ library/core/src/main/java/com/google/android/exoplayer2/source/ads/AdsMediaSource.java
@@ -119,7 +119,6 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
   private AdPlaybackState adPlaybackState;
   private MediaSource[][] adGroupMediaSources;
   private long[][] adDurationsUs;
-  private MediaSource.Listener listener;
 
   /**
    * Constructs a new source that inserts ads linearly with the content specified by {@code
@@ -204,11 +203,10 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
   }
 
   @Override
-  public void prepareSource(final ExoPlayer player, boolean isTopLevelSource, Listener listener) {
-    super.prepareSource(player, isTopLevelSource, listener);
+  public void prepareSourceInternal(final ExoPlayer player, boolean isTopLevelSource) {
+    super.prepareSourceInternal(player, isTopLevelSource);
     Assertions.checkArgument(isTopLevelSource);
     final ComponentListener componentListener = new ComponentListener();
-    this.listener = listener;
     this.componentListener = componentListener;
     prepareChildSource(new MediaPeriodId(/* periodIndex= */ 0), contentMediaSource);
     mainHandler.post(new Runnable() {
@@ -276,8 +274,8 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
   }
 
   @Override
-  public void releaseSource() {
-    super.releaseSource();
+  public void releaseSourceInternal() {
+    super.releaseSourceInternal();
     componentListener.release();
     componentListener = null;
     deferredMediaPeriodByAdMediaSource.clear();
@@ -286,7 +284,6 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
     adPlaybackState = null;
     adGroupMediaSources = new MediaSource[0][];
     adDurationsUs = new long[0][];
-    listener = null;
     mainHandler.post(new Runnable() {
       @Override
       public void run() {
@@ -350,7 +347,7 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
           adPlaybackState.adGroupCount == 0
               ? contentTimeline
               : new SinglePeriodAdTimeline(contentTimeline, adPlaybackState);
-      listener.onSourceInfoRefreshed(this, timeline, contentManifest);
+      refreshSourceInfo(timeline, contentManifest);
     }
   }